Refactored AuthorizationHelper to Service
