+++ src/Events/Controllers/EventsController.cs
@@ -4,14 +4,14 @@ using System.Linq;
 using System.Net;
 using System.Text;
 using System.Threading.Tasks;
+
 using Altinn.Common.AccessToken.Configuration;
-using Altinn.Common.PEP.Interfaces;
-using Altinn.Platform.Events.Authorization;
 using Altinn.Platform.Events.Configuration;
 using Altinn.Platform.Events.Exceptions;
 using Altinn.Platform.Events.Models;
 using Altinn.Platform.Events.Services.Interfaces;
 using Altinn.Platorm.Events.Extensions;
+
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
@@ -29,9 +29,9 @@ namespace Altinn.Platform.Events.Controllers
     {
         private readonly IEventsService _eventsService;
         private readonly IRegisterService _registerService;
+        private readonly IAuthorization _authorizationService;
         private readonly ILogger _logger;
         private readonly string _eventsBaseUri;
-        private readonly AuthorizationHelper _authorizationHelper;
         private readonly AccessTokenSettings _accessTokenSettings;
 
         /// <summary>
@@ -40,16 +40,16 @@ namespace Altinn.Platform.Events.Controllers
         public EventsController(
             IEventsService eventsService,
             IRegisterService registerService,
+            IAuthorization authorizationService,
             IOptions<GeneralSettings> settings,
             ILogger<EventsController> logger,
-            IPDP pdp,
             IOptions<AccessTokenSettings> accessTokenSettings)
         {
             _registerService = registerService;
             _logger = logger;
             _eventsService = eventsService;
             _eventsBaseUri = $"https://platform.{settings.Value.Hostname}";
-            _authorizationHelper = new AuthorizationHelper(pdp);
+            _authorizationService = authorizationService;
             _accessTokenSettings = accessTokenSettings.Value;
         }
 
@@ -211,7 +211,7 @@ namespace Altinn.Platform.Events.Controllers
 
             if (events.Count > 0)
             {
-                events = await _authorizationHelper.AuthorizeEvents(HttpContext.User, events);
+                events = await _authorizationService.AuthorizeEvents(HttpContext.User, events);
             }
 
             if (events.Count > 0)
